+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nanofish"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2024"
 description = "🐟 A lightweight, `no_std` HTTP client for embedded systems built on top of Embassy networking."
 documentation = "https://docs.rs/nanofish"
@@ -8,6 +8,10 @@ repository = "https://github.com/rttfd/nanofish"
 license = "MIT"
 keywords = ["no_std", "embedded", "async", "http_client"]
 authors = ["rttf <contact@rttf.dev>"]
+
+[features]
+default = []
+tls = ["dep:embedded-tls", "dep:rand_core", "dep:rand_chacha"]
 
 [dependencies]
 defmt = "1.0.1"
@@ -21,4 +25,7 @@ embassy-net = { version = "0.7.0", features = [
 ] }
 embassy-time = { version = "0.4.0", features = ["defmt"] }
 embedded-io-async = { version = "0.6.1", features = ["defmt-03"] }
+embedded-tls = { version = "0.17.0", default-features = false, optional = true }
 heapless = { version = "0.8.0", features = ["defmt-03"] }
+rand_chacha = { version = "0.3", default-features = false, optional = true }
+rand_core = { version = "0.6.4", optional = true }
